pipelines/pyproject.toml
[project]
name = "cfb-pipelines"
version = "0.1.0"
description = "Data pipelines for CFB Analytics using dlt"
requires-python = ">=3.10"
dependencies = [
    "dlt[duckdb]>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
